requirements.txt
whitenoise==4.1.2
gunicorn==19.9.0
Django==2.1.11
django-heroku==0.3.1
djangorestframework==3.9.3